What if we made like driving but it is smooth
It's like driving but a lot smoother. You know the texture of yogurt? That's how we drive now. So SMOOTH! You know butter? Yeah me too. AND WE. are going... to drive like that !!! YEHAWW!! i ahve been unable to test this because there is no robot to test my code on. IF IT DOESN"T WORK i wil CRY. IF IT DOES WORK i wil ALSO CRY. OUT OF JOY and PURE HATRED for the beast that I have created. I AM ANIMATING THINGS THAT SHOULD NOT BE ANIMATED. THIS IN UNNATURAL WE MADE THE ROCK THINK BY SHOCKING THEM. MODERN SCIENCE IS A WONDER!!!!! i am currently lost in the forest and I am unable to get out. JOY! WONDER! BE AMAZED AT WHAT I HAVE CREATED! the time for revolution is now. right now. I AM GOING TO VENEZUELA AS SOON AS I FIGURE OUT HOW TO LEAVE THESE WOODS. Yours truly, atticusc orne. conret..ncon cornet.s coATTICUS cornett